requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiomysql==0.2.0
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dotenv==0.9.9
ffmpeg-python==0.2.0
flower==2.0.1
frozenlist==1.6.0
future==1.0.0
greenlet==3.2.1
humanize==4.12.3
idna==3.10
kombu==5.5.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.0.0
six==1.17.0
SQLAlchemy==2.0.40
tornado==6.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
youtube-dl==2021.12.17
yt-dlp==2025.4.30
